FEATURE: Set mattermost status based on vacation status (#3)
Optionally (default on) set users' mattermost status emoji/text when they are on vacation. The emoji is customizable via the site setting. If the user is not on vacation, and their status has this specific emoji, their status will be cleared.

Optionally (default off) this job can clear ALL non-discourse-managed mattermost statuses